big fixes, first line handler, root block boundaries. (19/19 tests pass)
realized that since my indentationRegex was capturing the first char,
if it captured a parenthesis, the match was being registered as a
'Braces' boundary.
Thus, my tests should have been failing for cases where indentation was
immediately followed by braces, an "indented brace block".
By sheer (bad) luck, my code happens to get the right answer if the
indented-brace-block is on the first line, which happened to be the case
in all my tests
So I added a test with an indented-brace-block on a different line.

To fix the issue:
I changed my regex to use lookahead so it didn't capture the first char.
However, the first line now had to be handled separately because if it
had zero indentation, the match would have length zero, which causes the
regex matcher to get stuck in an infinite loop.
Thus, in both setIndentSequence and getBlockIterator, there is special
code for handling the first line indentation.

I also treat the entire text block as an indented block now, so there
is an additional indent boundary at the start and end. This fixes the
issues I was having with the rootBlock in constructBlockTree

Also had to update snapshots to account for root block boundaries, and
the change in the indentation regex