changed setIndentSequence() to getIndentationLength(), 20/20 tests pass
instead of using an initial pass to set the indentSequence, instead
getIndentationLength() is called while iterating through blocks, and
sets the indentSequence dynamically.

This consolidates the pass for setting the indentSequence, and the pass
for iterating through block boundaries, into a single pass. We also only
have to use all the regexes once